requirements.txt
streamlit>=1.28.0
google-generativeai>=0.3.0
requests>=2.31.0
streamlit-authenticator>=0.2.3
pydub>=0.25.1
speech-recognition>=3.10.0
gtts>=2.4.0
streamlit-webrtc>=0.47.0
python-dotenv>=1.0.0
pandas>=2.0.0
sqlite3
datetime
uuid
